requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
Django==1.11.7
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-nose==1.4.5
django-redis-cache==1.7.1
djangorestframework==3.7.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
nose==1.3.7
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
redis==2.10.6
requests==2.18.4
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
